Turn numalgos into child module of peer_did module
