+++ did_peer/src/lib.rs
@@ -1,6 +1,5 @@
 extern crate display_as_json;
 
 pub mod error;
-mod numalgos;
 pub mod peer_did;
-pub mod peer_did_resolver;
+pub mod resolver;
