--- did_peer/src/numalgos/mod.rs
@@ -1,2 +1,0 @@
-pub mod numalgo2;
-pub mod numalgo3;
