--- did_peer/src/numalgos/numalgo3/mod.rs
@@ -1,1 +1,0 @@
-mod generate;
