+++ did_peer/src/peer_did/numalgos/numalgo1/mod.rs
@@ -1,4 +1,4 @@
-use super::traits::Numalgo;
+use crate::peer_did::numalgos::Numalgo;
 
 #[derive(Clone, Copy, Default, Debug, PartialEq)]
 pub struct Numalgo1;
