+++ did_peer/src/peer_did/numalgos/numalgo2/mod.rs
@@ -1,22 +1,25 @@
-mod encoding;
-mod purpose;
-mod resolve;
-mod service_abbreviated;
-mod verification_method;
-
 use did_doc::schema::did_doc::DidDocument;
 use did_doc_sov::extra_fields::ExtraFieldsSov;
-pub use resolve::resolve_numalgo2;
+use did_parser::Did;
+use encoding::{append_encoded_key_segments, append_encoded_service_segment};
 use sha256::digest;
 
 use crate::{
     error::DidPeerError,
-    numalgos::numalgo2::encoding::{append_encoded_key_segments, append_encoded_service_segment},
     peer_did::{
-        numalgos::{numalgo2::Numalgo2, numalgo3::Numalgo3},
+        numalgos::{
+            numalgo2::resolve::resolve_numalgo2, numalgo3::Numalgo3, Numalgo, ResolvableNumalgo,
+        },
         FromDidDoc, PeerDid,
     },
+    resolver::options::PublicKeyEncoding,
 };
+
+mod encoding;
+mod purpose;
+pub mod resolve;
+mod service_abbreviated;
+mod verification_method;
 
 impl FromDidDoc for Numalgo2 {
     fn from_did_doc(
@@ -38,5 +41,22 @@ impl PeerDid<Numalgo2> {
         let numalgoless_id = self.did().id().chars().skip(2).collect::<String>();
         let numalgoless_id_hashed = digest(numalgoless_id);
         PeerDid::<Numalgo3>::parse(format!("did:peer:3.{}", numalgoless_id_hashed))
+    }
+}
+
+#[derive(Clone, Copy, Default, Debug, PartialEq)]
+pub struct Numalgo2;
+
+impl Numalgo for Numalgo2 {
+    const NUMALGO_CHAR: char = '2';
+}
+
+impl ResolvableNumalgo for Numalgo2 {
+    fn resolve(
+        &self,
+        did: &Did,
+        public_key_encoding: PublicKeyEncoding,
+    ) -> Result<DidDocument<ExtraFieldsSov>, DidPeerError> {
+        resolve_numalgo2(did, public_key_encoding).map(|builder| builder.build())
     }
 }
