--- did_peer/src/peer_did/numalgos/numalgo3.rs
@@ -1,8 +1,0 @@
-use super::traits::Numalgo;
-
-#[derive(Clone, Copy, Default, Debug, PartialEq)]
-pub struct Numalgo3;
-
-impl Numalgo for Numalgo3 {
-    const NUMALGO_CHAR: char = '3';
-}
