+++ did_peer/src/resolver/options.rs
