+++ did_peer/tests/resolve_positive.rs
@@ -2,7 +2,7 @@ mod fixtures;
 
 use did_doc::schema::did_doc::DidDocument;
 use did_doc_sov::extra_fields::ExtraFieldsSov;
-use did_peer::peer_did_resolver::{
+use did_peer::resolver::{
     options::{ExtraFieldsOptions, PublicKeyEncoding},
     resolver::PeerDidResolver,
 };
